Updated Payment flow

+++ src/main/java/com/orderms/entity/Order.java
@@ -33,7 +33,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long orderId;
 
-    private Long userId;  // Reference to UserMS
+    private Integer userId;  // Reference to UserMS
 
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL, orphanRemoval = true)
     @JsonManagedReference
